requirements.txt
alembic==1.16.1
blinker==1.9.0
click==8.1.7
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.0
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
GitPython==3.1.40
Sphinx==7.2.6
